# Optimize controller tests, fixes #307