This consolidates each test to one setup per context. For example, in master:
``` bash
$ bundle exec rspec spec/controllers/api/items_controller_spec.rb
[Coveralls] Set up the SimpleCov formatter.
[Coveralls] Using SimpleCov's default settings.
...........................................................

Finished in 3 minutes 31.8 seconds (files took 5.35 seconds to load)
59 examples, 0 failures
```
Now:
``` bash
$ bundle exec rspec spec/controllers/api/items_controller_spec.rb
[Coveralls] Set up the SimpleCov formatter.
[Coveralls] Using SimpleCov's default settings.
......................

Finished in 1 minute 11.24 seconds (files took 5.21 seconds to load)
22 examples, 0 failures
```